# port from enhanced-resolve to @rspack/resolver to solve memory leak with yarn pnp