--- a/packages/knip/README.md
+++ b/packages/knip/README.md
@@ -41,4 +41,4 @@ Special thanks to [the wonderful people who have contributed to Knip][8]!
 [5]: https://github.com/webpro-nl/knip/blob/main/.github/CONTRIBUTING.md
 [6]: https://knip.dev/sponsors
 [7]: https://www.youtube.com/watch?v=PE7h7KvQoUI&t=9s
-[8]: https://knip.dev/#-contributors
+[8]: https://knip.dev/#created-by-awesome-contributors