--- a/packages/knip/src/util/resolve.ts
+++ b/packages/knip/src/util/resolve.ts
@@ -1,23 +1,18 @@
-import fs from 'node:fs';
-import ER from 'enhanced-resolve';
+import { ResolverFactory } from '@rspack/resolver';
 import { DEFAULT_EXTENSIONS } from '../constants.js';
 import { timerify } from './Performance.js';
 import { toPosix } from './path.js';
 
-// @ts-ignore error TS2345 (not in latest): Argument of type 'typeof import("node:fs")' is not assignable to parameter of type 'BaseFileSystem'.
-const fileSystem = new ER.CachedInputFileSystem(fs, 9999999);
-
 export const createSyncResolver = (extensions: string[]) => {
-  const resolver = ER.create.sync({
-    fileSystem,
+  const resolver = new ResolverFactory({
     extensions,
     conditionNames: ['require', 'import', 'node', 'default'],
   });
 
   return function resolveSync(specifier: string, baseDir: string) {
     try {
-      const resolved = resolver({}, baseDir, specifier);
-      if (resolved) return toPosix(resolved);
+      const result = resolver.sync(baseDir, specifier);
+      if (result.path) return toPosix(result.path);
     } catch (_error) {}
   };
 };