@webpro 2025-05-13
Thanks for the PR, Matiss. Despite being in draft status I went ahead and did the same as I did in #1081:

## [Astro](https://github.com/withastro/astro)

Before

```
Name         size  min   max   median   sum  
-----------  ----  ----  ----  ------  ------
resolveSync  2192  0.02  6.26    0.08  296.10
resolveSync  2192  0.01  9.99    0.09  315.74
resolveSync  2192  0.01  9.98    0.09  310.10
```

After

```
Name         size  min   max   median   sum  
-----------  ----  ----  ----  ------  ------
resolveSync  2192  0.00  2.73    0.04  161.32
resolveSync  2192  0.00  0.80    0.04  142.58
resolveSync  2192  0.00  0.73    0.04  142.73
resolveSync  2192  0.00  8.05    0.04  153.72
```

## [Storybook](https://github.com/storybookjs/storybook)

Before

```
Name         size  min    max   median   sum  
-----------  ----  ----  -----  ------  ------
resolveSync  7600  0.01   8.72    0.05  653.81
resolveSync  7600  0.01  16.24    0.05  631.81
resolveSync  7600  0.01  16.03    0.05  639.96
resolveSync  7600  0.01  15.08    0.05  694.67
```

After

```
Name         size  min   max   median   sum  
-----------  ----  ----  ----  ------  ------
resolveSync  7600  0.00  0.74    0.01  305.43
resolveSync  7600  0.00  0.72    0.01  269.98
resolveSync  7600  0.00  0.68    0.01  270.80
```

I've just pushed a change to `main`:

- Includes `timerify` on `resolveSync` when used from `createCustomModuleResolver`
- Run `knip --performance-fn resolveSync` so we `timerify` and focus on only one function, for less performance overhead of `timerify` itself and a tiny table in the terminal output

Would love to merge either this or #1081.

Could you please rebase your changes on top of `main`?

If `oxc-resolver` turns out to be faster than `@rspack/resolver`, can you also please verify that `oxc-resolver` has the same characteristics memory-usage-wise in your project(s)?