@MatissJanis 2025-05-13
Thanks!

I'm still running knip in various permutation on our monorepo to get some more data, but it's taking some time because of the size of our repo (~3h per one run).

Here's the data I have so far on memory consumption with the patch applied:

**--isolated-workspaces:**
<img width="920" alt="Screenshot 2025-05-13 at 10 21 34" src="https://github.com/user-attachments/assets/ad33ebb9-a64b-4b8d-9123-a6e0bc364417" />

(the job times out eventually; looks like there is still some memory leak somewhere)

**NO workspace isolation:**
<img width="924" alt="Screenshot 2025-05-13 at 10 22 23" src="https://github.com/user-attachments/assets/88359fd3-33e0-4943-b165-a6bf1c366bff" />

(job is successful!)

For us **time** is not a constraint, but memory usage is. I'll try to run our project with #1081 patch too and see what's the memory consumption there (but it will take some time, so please be patient with me!). Will report my findings directly on that PR.