@webpro 2025-05-13
Sorry I forgot to remove the unused `createSyncResolver` export so just force-pushed a change to `main`

> --isolated-workspaces:
> 
> (the job times out eventually; looks like there is still some memory leak somewhere)

Just to be 100% sure, you not using a `dist`/latest version like v5.55.1 but from e52b129776f9e8dcf7ecc677378ccece95630739 or later, right?

Makes me think it's indeed Yarn PnP at least playing a role here. I guess there's no easy way to rule that out by not using PnP?

> please be patient with me 

Absolutely, no.. _rush_ here..