@MatissJanis 2025-05-13
> Just to be 100% sure, you not using a `dist`/latest version like v5.55.1 but from [e52b129](https://github.com/webpro-nl/knip/commit/e52b129776f9e8dcf7ecc677378ccece95630739) or later, right?

Yes, I am using the `process.exit(0)` patch as the latest change + the patch from this PR on top (rspack resolver).

> I guess there's no easy way to rule that out by not using PnP?

Unfortunately - no. Yarn PnP is deeply ingrained in our monorepo with no easy way to remove it. Though I would venture a guess that the issue is a specific memory leak in `enhanced-resolve`. It was recently patched in `rspack-resolver` [here](https://github.com/web-infra-dev/rspack-resolver/pull/69). I will report back how things looks with `osx-resolver`, but I'm quite hopeful that it will work (judging from the fact that they seem to support Yarn PnP - [here](https://github.com/oxc-project/oxc-resolver/issues/53)).